requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pillow>=10.2.0
pyyaml>=6.0.1
requests>=2.31.0
datasets>=2.18.0
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-multipart>=0.0.9
